Implemented angular material modals.

+++ src/app/modules/dashboard/dashboard-options/dashboard-options.component.ts
@@ -1,10 +1,10 @@
-import { Component, OnInit, Input, Output, ViewChild, EventEmitter } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { UserDashboard } from 'src/app/models/user-dashboard.model';
 import { DashboardDetailsModalComponent } from '../dashboard-details-modal/dashboard-details-modal.component';
-import { ModalComponent } from '../../display-elements/modal/modal.component';
 import { AuthService } from 'src/app/services/auth.service';
 import { DashboardService } from 'src/app/services/dashboard.service';
-import { MatDialog } from '@angular/material';
+import { MatDialog, MatDialogRef } from '@angular/material';
+import { ConfirmModalComponent } from '../../display-elements/confirm-modal/confirm-modal.component';
 
 @Component({
   selector: 'app-dashboard-options',
@@ -25,9 +25,6 @@ export class DashboardOptionsComponent implements OnInit {
   @Output() enterEditMode: EventEmitter<any>;
   @Output() leaveEditMode: EventEmitter<boolean>;
 
-  @ViewChild('dashboardDetailsModal') private dashboardDetailsModal: DashboardDetailsModalComponent;
-  @ViewChild('confirmDashboardDeletionModal') private dashboardDeletionModal: ModalComponent;
-
   constructor(private authSvc: AuthService, private dashSvc: DashboardService, private dialog: MatDialog) { 
     this.setDashboard = new EventEmitter();
     this.enterEditMode = new EventEmitter();
@@ -38,18 +35,18 @@ export class DashboardOptionsComponent implements OnInit {
   }
 
   setDashboardDetails(){
-    let detailRef = this.dialog.open(DashboardDetailsModalComponent, {
-      data: {
-        title: this.userDashboards[this.dashIndex].title,
-        description: (this.userDashboards[this.dashIndex].description ? this.userDashboards[this.dashIndex].description : '')
-      }
+    let modalRef: MatDialogRef<DashboardDetailsModalComponent> = this.dialog.open(DashboardDetailsModalComponent, {
+      
     });
 
-    detailRef.afterClosed().subscribe(result => {
-      this.userDashboards[this.dashIndex].title = result.title;
-      this.userDashboards[this.dashIndex].description = result.description;
-    });
+    modalRef.componentInstance.dashTitle = this.userDashboards[this.dashIndex].title;
+    modalRef.componentInstance.dashDescription = (this.userDashboards[this.dashIndex].description ? this.userDashboards[this.dashIndex].description : '');
 
+    modalRef.componentInstance.details.subscribe(details => {
+      this.userDashboards[this.dashIndex].title = details.title;
+      this.userDashboards[this.dashIndex].description = details.description;
+      modalRef.close();
+    });
   }
 
   createNewDashboard(){
@@ -64,27 +61,30 @@ export class DashboardOptionsComponent implements OnInit {
   }
 
   deleteDashboard(){
-    let deleteRef = this.dialog.open(ModalComponent, {
-      data: {
-        title: 'Delete Dashboard',
-        message: 'Are you sure you want to delete ' + this.userDashboards[this.dashIndex].title + ' from your dashboards?',
-        icons: [{icon: 'delete', classList: ''}],
-        buttons: [{title: 'Confirm', classList: 'btn btn-danger'}]
-      }
+    let modalRef: MatDialogRef<ConfirmModalComponent> = this.dialog.open(ConfirmModalComponent, {
+      
     });
 
-    deleteRef.afterClosed().subscribe(result => {if(result === 'confirm'){
-      this.leaveEditMode.emit(true);
+    modalRef.componentInstance.title = 'Delete Dashboard';
+    modalRef.componentInstance.message = 'Are you sure you want to delete ' + this.userDashboards[this.dashIndex].title + ' from your dashboards?';
+    modalRef.componentInstance.icons =  [{icon: 'delete', classList: ''}];
+    modalRef.componentInstance.buttons = [{title: 'Confirm', classList: 'btn btn-danger'}];
 
-      if(this.userDashboards[this.dashIndex].docId){
-        this.dashSvc.deleteDashboard(this.userDashboards[this.dashIndex].docId).subscribe(
-          (dashboard) => {this.deleteLocalDashboard();}
-        );
+    modalRef.componentInstance.btnClick.subscribe(btnClick => {
+      if(btnClick === 'confirm'){
+        this.leaveEditMode.emit(true);
+  
+        if(this.userDashboards[this.dashIndex].docId){
+          this.dashSvc.deleteDashboard(this.userDashboards[this.dashIndex].docId).subscribe(
+            (dashboard) => {this.deleteLocalDashboard();}
+          );
+        }
+        else{
+          this.deleteLocalDashboard();
+        }
       }
-      else{
-        this.deleteLocalDashboard();
-      }
-    }});
+      modalRef.close();
+    });
   }
 
   setDefault(): void{
